@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
 
     <Version>1.0.0</Version>
 
@@ -11,9 +11,11 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" Version="2.1.0" AllowExplicitVersion="True" />
-    <PackageReference Include="Microsoft.Extensions.Logging" Version="2.1.0" />
+
+    <!-- Explicit framework reference because this project builds a library to support APIs, not the actual APIs.-->
+    <FrameworkReference Include="Microsoft.AspNetCore.App" />
+
     <PackageReference Include="NEST" Version="5.6.*" />
-    <PackageReference Include="NSwag.AspNetCore" Version="13.2.*" />
+    <PackageReference Include="NSwag.AspNetCore" Version="13.11.*" />
   </ItemGroup>
 </Project>
